fixed skill info crash bug

+++ HappyTravel/Common/UserInfo.swift
@@ -13,7 +13,7 @@ import RealmSwift
 class Tally: Object {
     
     dynamic var tally:String?
-    dynamic var labelWidth:CGFloat = 0.0
+    dynamic var labelWidth:Float = 0.0
 }
 
 class PhotoUrl: Object {
@@ -147,7 +147,7 @@ class UserInfo: Object {
                     let string:NSString = tag
                     let options:NSStringDrawingOptions = [.UsesLineFragmentOrigin, .UsesFontLeading]
                     let rect = string.boundingRectWithSize(CGSizeMake(0, 20), options: options, attributes: [NSFontAttributeName : UIFont.systemFontOfSize(12)], context: nil)
-                    tally.labelWidth = rect.size.width + 24
+                    tally.labelWidth = Float(rect.size.width) + 24
                     businessTags.append(tally)
                 }
                 break
@@ -183,7 +183,7 @@ class UserInfo: Object {
                     let string:NSString = tag
                     let options:NSStringDrawingOptions = [.UsesLineFragmentOrigin, .UsesFontLeading]
                     let rect = string.boundingRectWithSize(CGSizeMake(0, 24), options: options, attributes: [NSFontAttributeName : UIFont.systemFontOfSize(12)], context: nil)
-                    tally.labelWidth = rect.size.width + 24
+                    tally.labelWidth = Float(rect.size.width) + 24
                     travalTags.append(tally)
                 }
                 break
